refactor: 라우터/인증/이벤트리스너 관심사 분리
