+++ src/router.js
@@ -1,81 +1,51 @@
+import { isAuthenticated } from "./auth";
+
 export default function Router() {
   const routes = {};
 
   /**
    * 경로 추가
    * @param {string} path
-   * @param {HTMLElement} element
+   * @param {function} handler
    */
-  const addRoute = (path, element) => {
-    routes[path] = element;
+  const addRoute = (path, handler) => {
+    routes[path] = handler;
   };
 
   /**
    * 화면 렌더링
+   * @param {function} route
+   */
+  const render = (route) => {
+    // 정상 접근 시 렌더링 && 비정상 URL 접근 시 404 화면
+    document.querySelector("#root").innerHTML = route();
+  };
+
+  /**
+   * 라우터 가드
    * @param {string} path
    */
-  const render = (path) => {
-    // 로그인 정보 체크
-    const isUser = JSON.parse(localStorage.getItem("user"));
-
-    if (isUser && path === "/login") {
+  const guard = (path) => {
+    if (isAuthenticated() && path === "/login") {
       // 로그인 사용자가 로그인 페이지에 접근한 경우 홈 화면으로 전환
       navigateTo("/");
-    } else if (!isUser && path === "/profile") {
+    } else if (!isAuthenticated() && path === "/profile") {
       // 비로그인 사용자가 프로필 페이지에 접근한 경우 로그인 페이지로 전환
       navigateTo("/login");
     } else {
       // 정상 접근 시 렌더링 && 비정상 URL 접근 시 404 화면
       const route = routes[path] || routes["/404"];
-      document.querySelector("#root").innerHTML = route();
+      render(route);
     }
   };
 
   const init = () => {
     // 현재 경로에 맞는 페이지 렌더링
-    render(window.location.pathname);
+    guard(window.location.pathname);
 
     // popstate(앞으로가기/뒤로가기) 라우터 처리
     window.addEventListener("popstate", () => {
-      render(window.location.pathname);
-    });
-
-    //  a 태그가 동적으로 생성되므로, 이벤트 위임(bubbling) 방법 사용
-    const root = document.getElementById("root");
-    root.addEventListener("click", (e) => {
-      if (e.target.id === "logout") {
-        // 로그아웃 시 사용자 정보 제거
-        localStorage.removeItem("user");
-      }
-      if (e.target.tagName === "A") {
-        // 클릭 시 페이지 이동
-        e.preventDefault();
-        navigateTo(e.target.pathname);
-      }
-    });
-
-    // 로그인/프로필 수정 시 사용자 정보 저장
-    const setUserInfo = () => {
-      const loginForm = {
-        username: document.getElementById("username")?.value || "", // 이메일 또는 전화번호(사용자 이름)
-        email: document.getElementById("email")?.value || "", // 이메일
-        bio: document.getElementById("bio")?.value || "", // 자기소개
-      };
-
-      localStorage.setItem("user", JSON.stringify(loginForm));
-    };
-    root.addEventListener("submit", (e) => {
-      if (e.target?.id === "login-form") {
-        // 로그인 정보 저장
-        e.preventDefault();
-        setUserInfo();
-        navigateTo("/profile");
-      } else if (e.target?.id === "profile-form") {
-        // 프로필 정보 수정
-        e.preventDefault();
-        setUserInfo();
-        alert("프로필 수정이 완료되었습니다 :)");
-      }
+      guard(window.location.pathname);
     });
   };
 
@@ -86,7 +56,7 @@ export default function Router() {
   const navigateTo = (path) => {
     if (window.location.pathname !== path) {
       history.pushState(null, "", path);
-      render(path);
+      guard(path);
     }
   };
 
@@ -95,5 +65,6 @@ export default function Router() {
     navigateTo,
     render,
     init,
+    guard,
   };
 }
